+++ src/main/java/com/example/sns_project/repository/GoodRepository.java
@@ -4,10 +4,8 @@ import com.example.sns_project.domain.entity.Good;
 import com.example.sns_project.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface GoodRepository extends JpaRepository<Good,Long> {
     Boolean findByUser(User user);
 
-    Long countByPostId(Long postId);
+    Integer countByPostId(Long postId);
 }
